+++ commands/import.go
@@ -9,9 +9,10 @@ import (
 	"github.com/bugcrowd/secrets/libsecrets"
 )
 
-// Export gets all values from the secrets
+// Import bulk adds data to a secrets scope
 func Import(c *cli.Context) {
-	scope, err := libsecrets.NewScope("default")
+	scopeName := c.String("scope")
+	scope, err := libsecrets.GetScope(scopeName)
 	if err != nil {
 		g.LogError(err)
 	}